+++ test/Serilog.Sinks.HttpTests/Extensions/Xunit/TheoryOnMasterBranch.cs
@@ -6,8 +6,8 @@ namespace Xunit
     public class TheoryOnMasterBranch : TheoryAttribute
     {
         public override string Skip { get; set; } =
-            Environment.GetEnvironmentVariable("APPVEYOR_REPO_BRANCH") == "master" &&
-            Environment.GetEnvironmentVariable("APPVEYOR_PULL_REQUEST_NUMBER") == null
+            Environment.GetEnvironmentVariable("APPVEYOR_REPO_BRANCH") == "master"
+            && Environment.GetEnvironmentVariable("APPVEYOR_PULL_REQUEST_NUMBER") == null
                 ? null
                 : "Only run test on master branch";
     }